website/requirements.txt
Flask
mysql-connector-python==8.0.23
werkzeug
python-magic
